[iOS] Web content process occasionally crashes under VisibleSelection::adjustPositionForEnd https://bugs.webkit.org/show_bug.cgi?id=225072 rdar://77159489
Reviewed by Darin Adler.

Source/WebCore:

Add a helper method to check whether or not the given `Position` is in a tree scope. See WebKit/ChangeLog for
more details.

Test: editing/selection/ios/clear-selection-while-moving-selection-handles.html

* dom/Position.h:
(WebCore::Position::isInTreeScope const):

Source/WebKit:

Make `rangeForPointInRootViewCoordinates` robust in the case where the original selection may have become
orphaned or null, by checking that the `targetNode` and existing `selectionStart` or `selectionEnd` are in a
tree scope before accessing `treeScope()`.

* WebProcess/WebPage/ios/WebPageIOS.mm:
(WebKit::rangeForPointInRootViewCoordinates):

LayoutTests:

Add a layout test that programmatically clears the selection in the middle of changing it by moving selection
handles.

* editing/selection/ios/clear-selection-while-moving-selection-handles-expected.txt: Added.
* editing/selection/ios/clear-selection-while-moving-selection-handles.html: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@276688 268f45cc-cd09-0410-ab3c-d52691b4dbfc