Make img elements perform source selection even when not in the document. https://bugs.webkit.org/show_bug.cgi?id=222801
Reviewed by Ryosuke Niwa.

LayoutTests/imported/w3c:

* web-platform-tests/html/semantics/embedded-content/the-img-element/source-media-outside-doc-expected.txt: Added.
* web-platform-tests/html/semantics/embedded-content/the-img-element/source-media-outside-doc.html: Added.

Source/WebCore:

The rules for image source selection in the HTML spec do not prevent
them from running when an <img> is not in the document.  So we update
HTMLImageElement to call selectImageSource() when inserted into and
removed from a <picture> without checking whether it's in a document.

Test: imported/w3c/web-platform-tests/html/semantics/embedded-content/the-img-element/source-media-outside-doc.html

* html/HTMLImageElement.cpp:
(WebCore::HTMLImageElement::~HTMLImageElement): Remove useless call to
setPictureElement, since all it does is clear a weak pointer.
(WebCore::HTMLImageElement::insertedIntoAncestor):
(WebCore::HTMLImageElement::removedFromAncestor):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@276697 268f45cc-cd09-0410-ab3c-d52691b4dbfc