Make sure we invalidate the iterator in StorageMap::clear() https://bugs.webkit.org/show_bug.cgi?id=225164
Reviewed by Alex Christensen.

I recently introduced StorageMap::clear() but failed to call invalidateIterator()
like we do in other StorageMap functions that modify the internal HashMap. Since
we cleared the HashMap, we don't want to be keeping a HashMap iterator around.
Looking at the code, I do not think this can cause crashes because the
`index >= length()` check would fail in StorageMap::key(unsigned). That said, it
is best not to keep a bad iterator around.

* storage/StorageMap.cpp:
(WebCore::StorageMap::clear):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@276737 268f45cc-cd09-0410-ab3c-d52691b4dbfc